filtro menos extras
