+++ src/app/app-routing.module.ts
@@ -1,16 +1,10 @@
 import { NgModule } from '@angular/core';
 import { PreloadAllModules, RouterModule, Routes } from '@angular/router';
 
-const routes: Routes = [
-  { path: '', redirectTo: 'home', pathMatch: 'full' },
-  { path: 'home', loadChildren: () => import('./home/home.module').then( m => m.HomePageModule)},
-  { path: 'resultado', loadChildren: './resultado/resultado.module#ResultadoPageModule' },
-  { path: 'peticion', loadChildren: './peticion/peticion.module#PeticionPageModule' },
-  { path: 'habitacion', loadChildren: './habitacion/habitacion.module#HabitacionPageModule' },
 
 const routes: Routes = [
-  { path: '', redirectTo: 'home', pathMatch: 'full' },
-  { path: 'home', loadChildren: () => import('./home/home.module').then( m => m.HomePageModule)},
+  { path: '', redirectTo: 'peticion', pathMatch: 'full' },
+  { path: 'peticion', loadChildren: () => import('./peticion/peticion.module').then( m => m.PeticionPageModule)},
   { path: 'resultado', loadChildren: './resultado/resultado.module#ResultadoPageModule' },
 
 ];
