+++ src/app/peticion/peticion.page.scss
@@ -1,0 +1,14 @@
+.imagen{
+    height:20%;
+    width:50%;
+        
+  }
+  .labelFooter{
+      font-size: 12.5px;
+      text-align: center;
+       color: #D1D1D1;
+  }
+  
+  .tamanoLetter{
+     font-size: 12px; 
+  }
